+++ data/people.json
@@ -1,1 +1,1 @@
-[{"id":258,"name":"Ivan","age":10,"parent_permission":false,"classroom":"4","type":"student"},{"id":442,"name":"Sally","age":5,"parent_permission":true,"classroom":"1","type":"student"},{"id":277,"name":"Uncle Tee","age":28,"specialization":"Music","type":"teacher"},{"id":313,"name":"TKSarso","age":20,"specialization":"Coding","type":"teacher"}]
+[{"id":258,"name":"Ivan","age":10,"parent_permission":false,"classroom":"4","type":"student"},{"id":442,"name":"Sally","age":5,"parent_permission":true,"classroom":"1","type":"student"},{"id":277,"name":"Uncle Tee","age":28,"specialization":"Music","type":"teacher"},{"id":313,"name":"TKSarso","age":20,"specialization":"Coding","type":"teacher"},{"id":579,"name":"Kojo","age":15,"parent_permission":true,"classroom":"6","type":"student"}]
